# enhanced_prompt_service/docs/enhanced_prompt_openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Enhanced Prompt Service API",
    "version": "1.3.0",
    "description": "## Enhanced Prompt Service\n\nA sophisticated conversational AI service that provides intelligent, context-aware responses for surveillance system queries. This service combines semantic search, large language models, and conversation memory to deliver enhanced user experiences.\n\n### Key Features:\n- **Semantic Search**: Vector-based search through surveillance events\n- **Conversational AI**: Context-aware responses with conversation memory\n- **Clip Integration**: Automatic video clip URL generation\n- **Follow-up Questions**: Intelligent suggestion of relevant next questions\n- **Proactive Insights**: Pattern detection and anomaly identification\n\n### Authentication:\nAll endpoints require a valid JWT token in the Authorization header:\n```\nAuthorization: Bearer YOUR_JWT_TOKEN\n```\n\n### Rate Limiting:\n- 100 requests per minute per user\n- 20 requests per 10 seconds burst limit",
    "contact": {
      "name": "Surveillance System Team",
      "email": "support@surveillance-system.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.surveillance-system.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/conversation/{conversation_id}/history": {
      "get": {
        "summary": "Get Conversation History",
        "description": "Get the message history for a specific conversation",
        "operationId": "get_conversation_history",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "description": "Conversation identifier",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum messages to return",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationHistory"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversation/{conversation_id}": {
      "delete": {
        "summary": "Delete Conversation",
        "description": "Delete a conversation and all its message history",
        "operationId": "delete_conversation",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "description": "Conversation identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Conversation deleted successfully"
                    },
                    "conversation_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Format: Bearer {token}"
      }
    },
    "schemas": {
...
      "ConversationHistory": {
        "type": "object",
        "properties": {
          "conversation_id": {
            "type": "string",
            "description": "Conversation identifier",
            "example": "conv_123e4567-e89b-12d3-a456-426614174000"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "List of messages in conversation"
          },
          "total_messages": {
            "type": "integer",
            "description": "Total number of messages",
            "example": 12
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Conversation creation timestamp",
            "example": "2025-06-18T09:15:00Z"
          },
          "last_activity": {
            "type": "string",
            "format": "date-time",
            "description": "Last activity timestamp",
            "example": "2025-06-18T10:30:02Z"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Message identifier",
            "example": "msg_456"
          },
          "role": {
            "type": "string",
            "enum": ["user", "assistant"],
            "description": "Message role"
          },
          "content": {
            "type": "string",
            "description": "Message content",
            "example": "Show me security incidents from today"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Message timestamp",
            "example": "2025-06-18T10:30:00Z"
          },
          "metadata": {
            "type": "object",
            "description": "Message metadata",
            "example": {
              "confidence_score": 0.92,
              "evidence_count": 5
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error description",
            "example": "Invalid request parameters"
          },
          "error_code": {
            "type": "string",
            "description": "Specific error code",
            "example": "INVALID_QUERY"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Error timestamp",
            "example": "2025-06-18T10:30:00Z"
          },
          "path": {
            "type": "string",
            "description": "Request path",
            "example": "/api/v1/conversation"
          },
          "correlation_id": {
            "type": "string",
            "description": "Request correlation ID",
            "example": "req_123456"
          }
        }
      }
    }
  },
...
}